+++ src/orders/MarketOrderImpl.java
@@ -1,3 +1,5 @@
+package orders;
+
 public class MarketOrderImpl extends MarketOrder{
 
     public MarketOrderImpl(double assetPrice, double tradeVolume) {
